requirements.txt
google-genai>=1.0.0
PyQt6>=6.6.0
pyaudio>=0.2.14
mss>=9.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
